Doctest errors fixed online

+++ pandas/core/arrays/sparse/array.py
@@ -1293,7 +1293,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [nan, nan, 1.0, 2.0]
         Fill: nan
         IntIndex
-        Indices: array([2, 3])
+        Indices: array([2, 3], dtype=int32)
 
         Using a SparseDtype, you can also change the fill value as well.
 
@@ -1367,7 +1367,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [10, 11, 12]
         Fill: 10
         IntIndex
-        Indices: array([1, 2])
+        Indices: array([1, 2], dtype=int32)
         """
         # this is used in apply.
         # We get hit since we're an "is_extension_type" but regular extension
